lobby-create: capitalize whitelist IDs
whitelist.tf doesn't care, and it propogates through to making the
headers look a little nicer when we display the ID. It would be even
better to have the headers do some smart lookup for how to prettify the
whitelist IDs, but for now this will do. The smarter prettification can
be done later, since we're already planning on (eventually) refactoring
the lobby create options (where the whitelist id -> pretty name map is).